Homework6 done

+++ src/main/java/ru/otus/java/basic/Animal/Cat.java
@@ -2,12 +2,14 @@ package ru.otus.java.basic.Animal;
 
 public class Cat extends Animal {
 
-    public Cat(String name, int speedRun, int endurance) {
+    public Cat(String name, double speedRun, int endurance) {
+
         super(name, speedRun, endurance);
     }
 
     @Override
     public void swim(int distance){
+
         System.out.println("Кошки не умеют плавать совсем!");
     }
 }
